'Solved' the 'stuck'-problem for Dijkstra
added a temporary solution as bug fix for Dijkstra not finding a route
to 'itself': go to random successor.
Dijkstra doesn't find a route to itself, so for a group with only one
value (for example victory) we get a 'stuck' situation.
In that case: go to a successor.
Future: combine with breadth first… (!)